refactor: move exceptions to subpackage
