--- eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/conduit/ConduitNotFoundException.java
@@ -1,7 +1,0 @@
-package com.github.twitch4j.eventsub.socket.conduit;
-
-public class ConduitNotFoundException extends Exception implements ConduitSocketPoolException {
-    ConduitNotFoundException(String conduitId, Exception cause) {
-        super("Could not find existing Conduit for pool with ID: " + conduitId, cause);
-    }
-}
