--- eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/conduit/ConduitSocketPoolException.java
@@ -1,3 +1,0 @@
-package com.github.twitch4j.eventsub.socket.conduit;
-
-public interface ConduitSocketPoolException {}
