--- eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/conduit/CreateConduitException.java
@@ -1,7 +1,0 @@
-package com.github.twitch4j.eventsub.socket.conduit;
-
-public class CreateConduitException extends Exception implements ConduitSocketPoolException {
-    CreateConduitException(Exception cause) {
-        super("Failed to create Conduit with helix for pool", cause);
-    }
-}
