+++ eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/conduit/TwitchConduitSocketPool.java
@@ -15,6 +15,11 @@ import com.github.twitch4j.eventsub.EventSubTransportMethod;
 import com.github.twitch4j.eventsub.condition.EventSubCondition;
 import com.github.twitch4j.eventsub.socket.IEventSubConduit;
 import com.github.twitch4j.eventsub.socket.TwitchEventSocket;
+import com.github.twitch4j.eventsub.socket.conduit.exceptions.ConduitNotFoundException;
+import com.github.twitch4j.eventsub.socket.conduit.exceptions.ConduitResizeException;
+import com.github.twitch4j.eventsub.socket.conduit.exceptions.CreateConduitException;
+import com.github.twitch4j.eventsub.socket.conduit.exceptions.ShardRegistrationException;
+import com.github.twitch4j.eventsub.socket.conduit.exceptions.ShardTimeoutException;
 import com.github.twitch4j.eventsub.socket.events.EventSocketWelcomedEvent;
 import com.github.twitch4j.eventsub.subscriptions.SubscriptionType;
 import com.github.twitch4j.helix.TwitchHelix;
