+++ eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/conduit/exceptions/ConduitSocketPoolException.java
@@ -1,0 +1,3 @@
+package com.github.twitch4j.eventsub.socket.conduit.exceptions;
+
+public interface ConduitSocketPoolException {}
